+++ update.sql
@@ -126,3 +126,11 @@ alter table if exists owner add column apartmentn int8
 alter table if exists owner_aud add column apartmentn int8
 alter table if exists owner add column apartmentn int8
 alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
+alter table if exists owner add column apartmentn int8
+alter table if exists owner_aud add column apartmentn int8
